+++ src/ray/core_worker/object_interface.h
@@ -4,7 +4,7 @@
 #include "common.h"
 #include "plasma/client.h"
 #include "ray/common/buffer.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/status.h"
 
 namespace ray {
