+++ src/ray/core_worker/task_interface.h
@@ -3,7 +3,7 @@
 
 #include "common.h"
 #include "ray/common/buffer.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/status.h"
 
 namespace ray {
