+++ src/ray/gcs/client.h
@@ -6,7 +6,7 @@
 
 #include "ray/gcs/asio.h"
 #include "ray/gcs/tables.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/status.h"
 #include "ray/util/logging.h"
