+++ src/ray/gcs/redis_module/ray_redis_module.cc
@@ -3,7 +3,7 @@
 
 #include "ray/common/common_protocol.h"
 #include "ray/gcs/format/gcs_generated.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/status.h"
 #include "ray/util/logging.h"
 #include "redis_string.h"
