+++ src/ray/object_manager/connection_pool.h
@@ -12,7 +12,7 @@
 #include <boost/asio/error.hpp>
 #include <boost/bind.hpp>
 
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/status.h"
 
 #include <mutex>
