+++ src/ray/object_manager/object_buffer_pool.h
@@ -12,7 +12,7 @@
 
 #include "plasma/client.h"
 
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/status.h"
 
 namespace ray {
