+++ src/ray/object_manager/object_directory.h
@@ -10,7 +10,7 @@
 #include "plasma/client.h"
 
 #include "ray/gcs/client.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/object_manager/format/object_manager_generated.h"
 #include "ray/status.h"
