+++ src/ray/object_manager/object_manager.h
@@ -16,7 +16,7 @@
 #include "plasma/client.h"
 
 #include "ray/common/client_connection.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/status.h"
 
 #include "ray/object_manager/connection_pool.h"
