+++ src/ray/object_manager/object_store_notification_manager.h
@@ -11,7 +11,7 @@
 
 #include "plasma/client.h"
 
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/status.h"
 
 #include "ray/object_manager/object_directory.h"
