+++ src/ray/raylet/actor_registration.h
@@ -4,7 +4,7 @@
 #include <unordered_map>
 
 #include "ray/gcs/format/gcs_generated.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/raylet/task.h"
 
 namespace ray {
