+++ src/ray/raylet/lineage_cache.h
@@ -8,7 +8,7 @@
 #include "ray/common/common_protocol.h"
 #include "ray/raylet/task.h"
 #include "ray/gcs/tables.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/status.h"
 // clang-format on
