+++ src/ray/raylet/monitor.h
@@ -5,7 +5,7 @@
 #include <unordered_set>
 
 #include "ray/gcs/client.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 
 namespace ray {
