+++ src/ray/raylet/node_manager.cc
@@ -5,7 +5,7 @@
 #include "ray/status.h"
 
 #include "ray/common/common_protocol.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/raylet/format/node_manager_generated.h"
 #include "ray/stats/stats.h"
