+++ src/ray/raylet/reconstruction_policy.h
@@ -8,7 +8,7 @@
 #include <boost/asio.hpp>
 
 #include "ray/gcs/tables.h"
-#include "ray/id.h"
+#include "ray/common/id.h"
 #include "ray/util/util.h"
 
 #include "ray/object_manager/object_directory.h"
